+++ History.md
@@ -5,6 +5,8 @@
 New features & improvements
 
 - Updated signature representation and added a customizable "Authorised signatory" label
+- Loosened bank "Account number" validation to support various international formats
+- Added "Account Holder" and "Sort Code / Sort Order" fields, which are now included in the PDF if provided.
 
 ## 2026-02-19, version 2.1.1